+++ crates/codexi/src/logic/codexi/mod.rs
@@ -1,6 +1,7 @@
 // src/logic/codexi/mod.rs
 
 mod error;
+mod import;
 mod init_data;
 mod migration_v1;
 mod migration_v2;